Test data for dials.plot_Fo_vs_Fc.
refmac_final.mtz is the data set discussed in the "DIALS for ED" paper
(https://doi.org/10.1107/S2059798318007726), which describes the
processing of an electron diffraction dataset from a lysozyme
nanocrystal. A similar dataset could be produced following the tutorial
https://dials.github.io/documentation/tutorials/dials_for_ed.html.

This data set may be used to investigate the effects of dynamic
diffraction and other error sources in ED. The program
dials.plot_Fo_vs_Fc produces an Fo vs Fc graph, which can be used
to assess the overall influence of dynamic scattering on the weak
reflections. See Figure 6 in https://doi.org/10.1107/S2059798317010348
for an example.