Add some minimal integrated data from two sweeps
Integrated data derived from two 1° wedges, each of ten 0.1° images.
These data were generated with xia2 and DIALS from the first ten images
of each of the first two sweeps of the DIALS small molecule tutorial
data set.

A README.md is included that explains the provenance of the data.