Miscellaneous regression files (#60)
* Files for testing https://github.com/dials/dials/issues/190, copied
from dials_regression and renamed

* Files for https://github.com/dials/dials/issues/191. The data were
originally at dials_regression/dials-191 and were in older experiment
list and pickle reflection file format. Here the relevant files were
updated to
- give new, more descriptive, file names
- remove local file paths from the experiments
- save as modern experiment lists
- save reflections in message pack format

* Add files for testing https://github.com/dials/dials/issues/554 to dials-data.

As for dials-191, take the chance to update to modern .expt and .refl format,
and remove paths to non-existent files on /dls/

* Add SIM_MX data used in various tests

* An MTZ from XDS processing used for Delta CC1/2 testing

* Files for reflection predictor tests

* Files for test of dials.goniometer_calibration

* Move files from nexus_test_data to dials-data. As with other moves,
the chance was taken to update to modern expt/refl format and delete
paths that won't exist on user machines.